Update in_rds_mysql_log.rb
sorry i mistyped the original. However, i could not get your plugin to work unless updating the input and so on. I wasn't sure if you were still working on this, since it was not in fluentd plugins so i forked yours and made changes relevant to my companies needs and with the update i made to yours.  Moving it solely for audit and error logs from my end. I deployed to fluentd plugs/rubygems to make it easier (gave you credit for the origin of the code).  I am also adding testing and travis ci, etc.  So, if you do not want to make these changes, please feel free to cancel it.  As i have the changes i need already onto my code and planning on changing the plugin even further.  I appreciate your quick reply, i wasn't sure if you were still managing the plugin.  Great work and let me know if you have further questions about my change or if you decide to close it.  thanks!